// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  ADMIN: 'ADMIN',
  EDITOR: 'EDITOR'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const BlogStatus = {
  PUBLISHED: 'PUBLISHED',
  DRAFT: 'DRAFT'
} as const

export type BlogStatus = (typeof BlogStatus)[keyof typeof BlogStatus]


export const CommentStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED',
  SPAM: 'SPAM'
} as const

export type CommentStatus = (typeof CommentStatus)[keyof typeof CommentStatus]


export const PromoPlacement = {
  HOME: 'HOME',
  ABOUT: 'ABOUT',
  CONTACT: 'CONTACT',
  BLOG_POST: 'BLOG_POST'
} as const

export type PromoPlacement = (typeof PromoPlacement)[keyof typeof PromoPlacement]


export const ProjectCategory = {
  Design: 'Design',
  Web: 'Web',
  Mobile: 'Mobile',
  Writings: 'Writings',
  Startups: 'Startups',
  Software: 'Software',
  Other: 'Other'
} as const

export type ProjectCategory = (typeof ProjectCategory)[keyof typeof ProjectCategory]


export const AwardKind = {
  AWARD: 'AWARD',
  CERTIFICATE: 'CERTIFICATE'
} as const

export type AwardKind = (typeof AwardKind)[keyof typeof AwardKind]


export const ContactStatus = {
  NEW: 'NEW',
  READ: 'READ',
  REPLIED: 'REPLIED',
  ARCHIVED: 'ARCHIVED'
} as const

export type ContactStatus = (typeof ContactStatus)[keyof typeof ContactStatus]